# packages/api-node/openapi.yaml
openapi: 3.0.3
info:
  title: User Profile App - API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
...
  /api/profiles/{id}/follow:
    post:
      summary: Follow a user
      tags: [Profiles]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Followed }
  /api/profiles/{id}/unfollow:
    post:
      summary: Unfollow a user
      tags: [Profiles]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Unfollowed }
  /api/profiles/{id}/following:
    get:
      summary: List following of a profile
      tags: [Profiles]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/profiles/{id}/followers:
    get:
      summary: List followers of a profile
      tags: [Profiles]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...